Bank Lending
This project is basically focused for the banking sector. This  project helps us to find out the insights of large dataset and to understand the procedures and techniques required for removing the unwanted columns replacing the impure data and with meaningful data values.
With the help of python it is very interesting and smooth to do the implementation.